requirements.txt
numpy
flask
tensorflow
keras
gunicorn
pillow